dom/base/test/test_bug813919.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=813919
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 813919</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=813919">Mozilla Bug 813919</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 813919 **/

  function testDataNode(dataNode) {
    var div = document.createElement("div");
    div.appendChild(dataNode);
    var span = document.createElement("span");
    div.appendChild(span);
    var r = document.createRange();
    r.setStart(dataNode, 0);
    r.setEnd(div, div.childNodes.length);
    r.deleteContents();
    ok(r.collapsed, "Range should be collapsed!");
    is(r.startContainer, div, "startContainer should be div.");
    is(r.startOffset, div.childNodes.length,
       "Range should be collaped to the end of the div element.");
  }

  testDataNode(document.createProcessingInstruction("x", "x"));
  testDataNode(document.createComment("x"));
  testDataNode(document.createTextNode("x"));


</script>
</pre>
</body>
</html>
